Feat: Add second class excercise of day 3 - HTML
In class3-2.html we can see how to use the <audio> and <video> tags.
You're going to need a audio file in one of the following fomats:
.mp3, .wav, .ogg, .au
For the video you're going to need a video in one of the following formats:
.ogm, .ogv, .ogg
Also change the src in the video tag, write the route of your file, in the case of the audio tag a song
has been added.